Canadian Immigration Dataset.csv
The dataset provides the latest information available publicly on IRCC website. It tells what immigration programs are being targeted, how many people are being invited and what is the score cutoff for each rounds of invitation. A lower CRS Score means that more people will be happy and immigrants are constantly on a lookout for a trend where cores would fall. Based on historic data, you can predict whether the scores are going up or down in 2021.